pyproject.toml
[project]
name = "steam-games-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gamerec",
    "pyyaml>=6.0.2",
]

[dependency-groups]
duckdb = [
    "duckdb>=1.3.2"
]
bigquery = [
    "google-cloud-bigquery>=3.36.0"
]
scraping-base = [
    "bs4>=0.0.2",
    "google-cloud-firestore>=2.21.0",
    "polars>=1.30.0",
    "requests>=2.32.4",
    "pyarrow>=21.0.0",
]
orchestration-base = [
    "apache-airflow==3.0.4"
]
dbt-duckdb = [
    { include-group = "duckdb" },
    "dbt-duckdb>=1.9.3",
    "dbt-postgres>=1.9.1",
]
dbt-bigquery = [
    { include-group = "bigquery"},
    "dbt-bigquery>=1.10.1"
]
scraping-duckdb = [
    { include-group = "scraping-base" },
    { include-group = "duckdb" },
    "psycopg[binary]>=3.2.9",
]
scraping-bigquery = [
    { include-group = "scraping-base" },
    { include-group = "bigquery" }
]
orchestration-duckdb = [
    { include-group = "orchestration-base" },
    { include-group = "duckdb" }
]
orchestration-bigquery = [
    { include-group = "orchestration-base" },
    { include-group = "bigquery" }
]
notebooks = [
    "boto3>=1.39.11",
    "dotenv>=0.9.9",
    "duckdb>=1.3.2",
    "fbgemm-gpu>=1.2.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-text-splitters>=0.3.11",
    "matplotlib>=3.9.2",
    "mlflow @ git+https://github.com/mlflow/mlflow@master",
    "notebook>=7.0.7",
    "openai>=1.109.1",
    "polars[pyarrow]>=1.30.0",
    "pyarrow>=20.0.0",
    "s3fs>=2025.7.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.11.0",
    "torch>=2.7.1",
    "torchmetrics>=1.0.3",
    "torchrec>=1.2.0",
]

[tool.uv.sources]
gamerec = { workspace = true }

[tool.uv.workspace]
members = [
    "gamerec"
]
